Title: 최솟값 3개, Time: 91ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 231221/최솟값 3개/top-3-smallest-number.cpp
@@ -3,7 +3,7 @@
 #define ll long long
 using namespace std;
 //func
-int run();
+ll run();
 void init();
 
 //var
